backend/requirements.txt
colorama
appdirs==1.4.3
gevent
greenlet
packaging==16.8
pyparsing==2.2.0
PyJWT==1.5.3
PyYAML
six==1.10.0
Werkzeug==0.12.1
PyNaCl==1.2.1
